refactor code fort icon font asworme and Move files to Storage Folder
