--- src/app/about/about.component.html
@@ -1,58 +1,0 @@
-<div class="about-card">
-  <h1>About</h1>
-  <div class="about-content">
-    <p>
-      Chef’s Compass is a recipe finding app for you, the user. It helps you find a recipe based on your hunger-tackling
-      desires and any ingredients stocked in your kitchen. Not sure how to use them effectively? Let Chef’s Compass show
-      you the way.
-    </p>
-  </div>
-</div>
-
-<div class="cards">
-  <div class="arturo-card">
-    <img src="/assets/arturo.jpg" alt="arturo" />
-    <div class="profile">
-      <div class="text">
-        <h2>Arturo Pimentel</h2>
-        <h4>Web Developer</h4>
-      </div>
-      <div class="links">
-        <a href="https://github.com/ArPimentel" target="_blank"><img src="/assets/git.svg" alt="git" /></a>
-        <a href="https://www.linkedin.com/in/arturo-pimentel-developpeur-web/" target="_blank">
-          <img src="/assets/li.svg" alt="linked" />
-        </a>
-      </div>
-    </div>
-  </div>
-  <div class="tyler-card">
-    <img src="/assets/tyler.jpg" alt="tyler" />
-    <div class="profile">
-      <div class="text">
-        <h2>Tyler Durnion</h2>
-        <h4>Web Developer</h4>
-      </div>
-      <div class="links">
-        <a href="https://github.com/td6904" target="_blank"><img src="/assets/git.svg" alt="git" /></a>
-        <a href="https://www.linkedin.com/in/tylerdurnion/" target="_blank">
-          <img src="/assets/li.svg" alt="linked" />
-        </a>
-      </div>
-    </div>
-  </div>
-  <div class="remy-card">
-    <img src="/assets/remy.jpg" alt="arturo" />
-    <div class="profile">
-      <div class="text">
-        <h2>Remy</h2>
-        <h4>Web Developer</h4>
-      </div>
-      <div class="links">
-        <a href="https://github.com/ArPimentel" target="_blank"><img src="/assets/git.svg" alt="git" /></a>
-        <a href="https://www.linkedin.com/in/arturo-pimentel-developpeur-web/" target="_blank">
-          <img src="/assets/li.svg" alt="linked" />
-        </a>
-      </div>
-    </div>
-  </div>
-</div>
