--- src/app/about/about.component.scss
@@ -1,105 +1,0 @@
-
-body {
-    font-family: Arial, sans-serif;
-    background-color:#e9e9e9;
-    margin: 0;
-    padding: 0;
-  }
-
-  .cards {
-    display: flex;
-    justify-content: center;
-    gap: 10%;
-    border: 1px solid #e8e8e8f0;
-    margin-top: 50px;
-    flex-wrap: wrap;
-    padding-bottom: 50px;
-     
-  }
-  
-  .tyler-card,
-  .arturo-card,
-  .remy-card {
-    border-radius: 50%;
-    width: 50%; 
-    height: auto;
-    background-color: #fff;
-    border-radius: 10px;
-    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
-    padding: 20px;
-    text-align: center;
-    transition: transform 0.3s ease-in-out;
-    flex: 1 1 300px; 
-    margin: 10px; 
-    max-width: 300px;
-  }
-
-  .cards .tyler-card:hover,
-  .arturo-card:hover,
-  .remy-card:hover {
-    transform: translatex(-10px);
-    background-color: rgb(255, 136, 0); 
-  }
-  
-  .profile {
-    display: flex;
-    justify-content: space-between;
-    margin: 20px;
-  }
-  
-  h2 {
-    font-size: 20px;
-    margin-bottom: 5px;
-  }
-  
-  h4 {
-    font-size: 16px;
-    color: #888;
-    margin: 0;
-  }
-  
-  .links {
-    width: 30px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-  }
-  
-  .links a {
-    display: inline-block;
-    margin-right: 10px;
-  }
-  .cards img {
-    width: 100%; 
-    height: auto; 
-  }
-  
-  .links img {
-    width: 30px;
-  }
-  
-  a {
-    color: #333;
-    text-decoration: none;
-  }
-  
-  a:hover {
-    color: #666;
-  }
-  
-  .about-card {
-    background-color: #f2f2f2;
-    padding: 20px;
-    border-radius: 10px;
-  }
-  
-  .about-card h1 {
-    font-size: 24px;
-    margin-bottom: 10px;
-  }
-  
-  .about-content p {
-    font-size: 14px;
-    line-height: 1.5;
-    color: #666666;
-  }
