+++ src/app/component/error-modal/error-modal.component.scss
