+++ src/app/component/footer/footer.component.html
@@ -1,0 +1,27 @@
+<!-- Footer -->
+
+<footer>
+  <div class="content">
+    <div class="footer">
+      <h3>Web application</h3>
+      <p>© 2023 Chef's Compass. All rights reserved</p>
+    </div>
+      <div class="about">
+      <a [routerLink]="['/about']" class="about_logo">
+          <fa-icon [icon]="faPeopleGroup" size="xl"></fa-icon>
+      </a>
+
+      <a href="" class="about_logo">
+        <fa-icon [icon]="faGear" size="xl"></fa-icon>
+      </a>
+
+      <a href="" class="about_logo">
+        <fa-icon [icon]="faFacebook" size="xl"></fa-icon>
+      </a>
+
+      <a href="" class="about_logo">
+        <fa-icon [icon]="faInstagram" size="xl"></fa-icon>
+      </a>
+    </div>
+  </div>
+</footer>
