+++ src/app/component/footer/footer.component.scss
@@ -1,3 +1,4 @@
+@use '../../style/variables' as *;
 * {
     margin: 0;
   }
@@ -12,15 +13,7 @@
     justify-items: center;
 
   }
-  .logo {
 
-    margin: 0 1vh;
-   
-    cursor: pointer;
-    
-    fill: black;
-    
-  }
 
   .about {
     grid-area: 1 / 3 / 2 / 4;
@@ -28,9 +21,18 @@
     display: flex;
     margin-right: 4vh;
     justify-self: end;
+    &_logo {
 
+      margin: 0 1vh;
+      cursor: pointer;
+      fill: black;
+      &:hover {
+        scale: 1.2;
+        color: $primaryColor;
+      }
+    }
   }
-  
+
   .footer {
     grid-area: 1 / 2 / 2 / 3;
     display: flex;
@@ -48,4 +50,4 @@
   }
 
 
-  
+
