+++ src/app/component/footer/footer.component.spec.ts
