+++ src/app/component/navbar/navbar.component.html
@@ -1,7 +1,7 @@
 <header>
   <div class="chefs-compass">
     <a routerLink="/recipes" class="compass-link">
-      <img class="nav-logo" src="/assets/main-logo.svg" alt="home-link-picture" />
+      <img class="nav-logo" src="/assets/icon/NavBar/main-logo.svg" alt="home-link-picture" />
       <p class="slogan">Global flavours from your kitchen</p>
     </a>
   </div>
@@ -11,7 +11,7 @@
   <div style="display: none">
     <div class="search-module">
       <input type="text" placeholder="Dishes..." />
-      <button type="submit"><img src="/assets/mag-glass.png" /></button>
+      <button type="submit"><img src="/assets/icon/NavBar/mag-glass.png" /></button>
       <div class="mini-burger">
         <div class="line"></div>
         <div class="line"></div>
@@ -24,25 +24,25 @@
     <ul>
       <li>
         <a>
-          <img src="/assets/community-button.png" alt="nav-logo" />
+          <img src="/assets/icon/NavBar/community-button.png" alt="nav-logo" />
           <p>Community</p>
         </a>
       </li>
       <li>
         <a routerLink="favourites">
-          <img src="/assets/favorite-button.png" alt="nav-logo" />
+          <img src="/assets/icon/NavBar/favorite-button.png" alt="nav-logo" />
           <p>Favourites</p>
         </a>
       </li>
       <li>
         <a [routerLink]="['/add-recipe']" rel="noopener" title="About" class="about-link" class="logo">
-          <img src="/assets/add-button.png" alt="nav-logo" />
+          <img src="/assets/icon/NavBar/add-button.png" alt="nav-logo" />
           <p>Add Recipes</p>
         </a>
       </li>
       <li>
         <a routerLink="/recipes">
-          <img src="/assets/recipes-button.png" alt="nav-logo" />
+          <img src="/assets/icon/NavBar/recipes-button.png" alt="nav-logo" />
           <p>All Recipes</p>
         </a>
       </li>
