+++ src/app/component/success-modal/success-modal.component.scss
