+++ src/app/pages/about/about.component.html
@@ -16,12 +16,12 @@
             <h3>Web Developer</h3>
             <a [href]="team.getGithub()">
                 <p>Github</p>
-                <img src="../../../assets/icon/github.png" alt="">
+                <fa-icon [icon]="faGithub"></fa-icon>
                 </a>
-    
+
             <a [href]="team.getLinkedin()">
                 <p>Linkedin</p>
-                <img src="../../../assets/icon/linkedin.png" alt="">
+              <fa-icon [icon]="faLinkedin"></fa-icon>
             </a>
         </div>
     </li>
