+++ src/app/pages/about/about.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { Teams } from 'src/app/models/Team.model';
 import { TeamService } from 'src/app/services/Team/team.service';
+import { faGithub,faLinkedin } from "@fortawesome/free-brands-svg-icons";
 
 @Component({
   selector: 'app-about',
@@ -9,10 +10,12 @@ import { TeamService } from 'src/app/services/Team/team.service';
 })
 export class AboutComponent {
   teams:Teams[]
-  
-  constructor(private teamService: TeamService) { 
+
+  constructor(private teamService: TeamService) {
     this.teams = this.teamService.getAllTeams()
   }
+  faGithub = faGithub;
+  faLinkedin = faLinkedin;
 
   }
