+++ src/app/pages/homepage/homepage.component.html
@@ -1,6 +1,6 @@
 <div class="page-container">
   <div class="contents">
-    <img src="../../assets/main-logo.png" alt="">
+    <img src="/assets/logo/main-logo.png" alt="">
     <a class="button" routerLink="recipes"> Welcome to Chef's Compass</a>
 </div>
 </div>
