+++ src/app/pages/homepage/homepage.component.scss
@@ -1,6 +1,6 @@
-@use '../style/_mixin.scss' as *;
-@use '../style/_animation.scss' as *;
-@use '../style/_variables.scss' as *;
+@use '../../style/mixin' as *;
+@use '../../style/animation' as *;
+@use '../../style/variables' as *;
 
 .page-container {
     height: 100%; /* Make sure the body takes up the full height of the viewport */
@@ -13,16 +13,16 @@
 
   .contents {
     position: relative;
-    
+
     height: 87vh;
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
     margin: 0;
- 
+
     & .button{
-  
+
       margin-top: 1rem;
       margin-left: 5rem;
       padding: 1.5rem 1rem;
@@ -33,12 +33,12 @@
     }
 
   }
-  
+
   :host {
     display: block;
     position: relative;
   }
-  
+
   :host::before {
     content: "";
     position: absolute;
@@ -46,14 +46,14 @@
     left: 0;
     width: 100%;
     height: 100%;
-    background-image: url("/assets/background.png");
+    background-image: url("/assets/backgroundPage/background.png");
     background-repeat: no-repeat;
     background-size: cover;
     z-index: -1;
   }
-  
+
   .main-logo {
-    
+
     max-width: 100%;
     max-height: 100%;
   }
