+++ src/app/pages/recipe-form/recipe-form.component.html
@@ -4,7 +4,7 @@
 >
   <div class="container">
     <div class="video-container">
-      <video src="assets/video-add-recipe.mp4" autoplay loop muted></video>
+      <video src="assets/video/video-add-recipe.mp4" autoplay loop muted></video>
       <div class="video-overlay">
         <h2 class="video-title">Get Ready to Spice Things Up!</h2>
         <h3 class="video-subtitle">Add Your Recipe to Chef's Compass!</h3>
