+++ src/app/pages/single-recipe/single-recipe.component.scss
@@ -6,7 +6,7 @@
     display: block;
     position: relative;
   }
-  
+
   :host::before {
     content: "";
     position: absolute;
@@ -14,11 +14,11 @@
     left: 0;
     width: 100%;
     height: 100%;
-    background-image: url("/assets/background.png");
+    background-image: url("/assets/backgroundPage/background.png");
     background-repeat: no-repeat;
     background-size: cover;
     filter: opacity(0.2);
- 
+
     z-index: -1;
   }
 
@@ -28,11 +28,11 @@
     text-align: center;
     padding: 5vh;
     color: $primaryColor;
-    
+
 }
 
 .recipe{
-    margin:$marginPage; 
+    margin:$marginPage;
     position: relative;
     z-index: 1;
     &_fav{
@@ -48,10 +48,10 @@
         color: $primaryColor;
         }
 
-    }   
+    }
     &_information{
     display: flex;
-    align-items:flex-start; 
+    align-items:flex-start;
     justify-content: space-between;
 
     & h2{
@@ -71,7 +71,7 @@
         position: relative;
         bottom: 1vh;
         border-radius: 50px;
-     
+
     }
 
     }
@@ -79,8 +79,8 @@
         display: flex;
         justify-content: space-between;
         margin-top:2vh ;
-        
-       
+
+
         &--origin{
             display: flex;
             justify-content: flex-start;
@@ -97,40 +97,40 @@
         display: flex;
         justify-content: space-between;
 
-    
+
         &--send{
             display: flex;
             flex-direction: column;
             align-items: flex-start;
-           
+
         & textarea{
             resize: none;
             border-radius:25px ;
             border: 1px solid $primaryColor;
             padding: 10px;
             &:focus{
-                background-color: lighten($color: $primaryColor, $amount: 10);    
+                background-color: lighten($color: $primaryColor, $amount: 10);
                 color: white;
                 outline: none;
             }
-           
-        }    
+
+        }
         & button{
           @include button;
-            
-            &:hover{ 
+
+            &:hover{
                 @include buttonHover;
               }
         }
         &--flex{
-            
+
                 display: flex;
                 flex-direction: column;
                 align-items: center;
-        
-            
+
+
         }
-        
+
         }
         &--msg p {
             max-width: 400px;
@@ -143,7 +143,7 @@
 
         }
     }
-    
+
 }
 .butn{display: none;}
 .rating > label{ position: relative;}
